length optimizer: relax the dependencies of exterior triangles
Childless "exterior" triangles can be completed in any order.
If the triangle has children, make sure to complete the edge
opposite of the final vertex early (not as the last edge).
Otherwise the greedy optimizer might move the completion of
that edge last, causing two (or more) triangles to be
constructed at the same side of the edge at the same time.

This usually shaves off some additional percents of the
total route length.