Improve performance of agentOrder.improveEdgeOrderMore
Do not recompute the length of the path in each iteration of improveEdgeOrderMore.
Instead, in each iteration compute only lengths of (at most) 3 removed edges,
 and (at most) 3 added edges and use it to update the original path length.
This brings down complexity of the algorithm significantly.
For my test of 150 portals it brings time down from over 1 minute to 3 seconds.
It's now feasible to run the improveEdgeOrderMore on every candidate edge order,
and choose the "best" edge order more accurately.